pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "inflamm_debate_fm"
version = "0.0.1"
description = "Revisiting the mouse-human inflammation response debate using foundation models"
authors = [
  { name = "Ahmed Sallam" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "anndata>=0.10.0",
    "flytekit>=1.10.0",
    "GEOparse>=1.2.0",
    "gseapy>=1.1.10",
    "ipykernel>=7.1.0",
    "loguru",
    "matplotlib>=3.10.7",
    "mygene>=3.2.2",
    "numpy>=1.24.0",
    "pandas>=2.3.3",
    "pip",
    "pre-commit>=4.3.0",
    "python-dotenv",
    "requests>=2.31.0",
    "ruff",
    "scanpy>=1.11.5",
    "scikit-dimension>=0.3.4",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tomli>=2.0.0; python_version < '3.11'",
    "torch>=2.0.0",
    "torch-geometric>=2.0.0",
    "tqdm",
    "typer",
    "wandb>=0.22.2",
    "xgboost>=3.1.1",
]
requires-python = ">=3.10"


[tool.ruff]
line-length = 99
src = ["inflamm_debate_fm"]
include = ["pyproject.toml", "inflamm_debate_fm/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["inflamm_debate_fm"]
force-sort-within-sections = true

